GC1M0D0 - Bellevue Blackout Challenge (United States)
Edit 2025-07-04: Cleaned up the listing.  Clarified that there are four challenge caches and one adventure lab bonus.
Okay, I had some free time to take a look at the Project-GC script and made a few changes:
* **Excluded Challenge and Bonus caches** - I spent most of the morning looking at the API and other checker examples to do this programmatically, including limiting it to Unknown Caches, and it's kind of involved.  Turns out there are only **five** that affect this, so I created a hard-coded list for the four challenge caches (GC24KA3, GCB75J8, GCAHV69 and GCAVD5Z) and one Adventure Lab bonus cache (GCB15TC).   **If you see more of these pop up, please let me know.**
* **Added a missing tag** so the checker did not erroneously report caches *you own* as to find.
* Fixed a bug where the checker would fail because were too many caches retrieved.

For the latter -- and this is primarily a note to myself -- The checker starts by calling *PGC.GetOldestCaches* to retrieve caches in a county/region(state)/country entity.  It excludes archived and disabled.  Because *[King County has a lot of caches](https://www.geocaching.com/geocache/GC1A012)*, and the checker seeks to not bog down the Project-GC server, it does multiple queries limiting to particular cache types, difficulty and terrain to keep the results to fewer than 1000 caches on each call.

The "Traditional caches of Difficulty 1.5 and terrain rating of 1.0 or 1.5" call was returning just over that limit, which would cause the checker to always return a false result that was largely unnoticed because this challenge *is hard*.   I broke this down into two queries.   Currently:

* Traditional D1.5, T1.0 = 111
* Traditional D1.5, T1.5 = 910
* Traditional D>1.5, T 1.0 or 1.5 = 938
* Traditional D != 1.5, T 2 = 401
* Traditional D != 1.5, T > 2 = 733
* Non-traditional T 1.0 or 1.5 = 847
* Non-traditional T > 1.5 = 741